do you really want to do this confirmation (untested)

+++ cleaner_cli/clean.py
@@ -32,27 +32,31 @@ def clean():
     #        raise ValueError('Folder does not exist.')
     filetypes = options.filetypes.split(',')
 
-    for ft in filetypes:
-        files = glob.glob('*.{t}'.format(t = ft))
-        if len(files)>0:
-            # if there isn't already a folder named `ft`, make one
-            if not exists(join(folder, ft)):
-                os.mkdir(join(folder, ft))
-            # move each file from folder to folder/`ft`
-            for f in files:
-                filename = basename(f)
-                # if the file is not exempt
-                if splitext(basename(filename))[0] not in options.exemptions.split(','):
-                    if not exists(join(folder, ft, filename)):
-                        os.rename(f, join(folder, ft, filename))
-                    else:
-                        # if a file already existed, should it be replaced?
-                        user_response = input('File {f} already exists in folder. Overwrite? [Y/n]'.format(f = filename))
-                        if user_response == 'Y':
-                            os.remove(join(folder, ft, filename))
+    continue_answer = input('Continue cleaning files {ft} in folder {f} (Y)'.format(ft = filetypes, f = pwd() ))
+    if continue_anwer == 'Y':
+        for ft in filetypes:
+            files = glob.glob('*.{t}'.format(t = ft))
+            if len(files)>0:
+                # if there isn't already a folder named `ft`, make one
+                if not exists(join(folder, ft)):
+                    os.mkdir(join(folder, ft))
+                # move each file from folder to folder/`ft`
+                for f in files:
+                    filename = basename(f)
+                    # if the file is not exempt
+                    if splitext(basename(filename))[0] not in options.exemptions.split(','):
+                        if not exists(join(folder, ft, filename)):
                             os.rename(f, join(folder, ft, filename))
                         else:
-                            pass
-                else:
-                    pass
-    print("Files of types {fts} have been organized in folders.".format(fts = options.filetypes))
+                            # if a file already existed, should it be replaced?
+                            user_response = input('File {f} already exists in folder. Overwrite? [Y/n]'.format(f = filename))
+                            if user_response == 'Y':
+                                os.remove(join(folder, ft, filename))
+                                os.rename(f, join(folder, ft, filename))
+                            else:
+                                pass
+                    else:
+                        pass
+        print("Files of types {fts} have been organized in folders.".format(fts = options.filetypes))
+    else:
+        print("Clean aborted")
